+++ identity/auth-vue/strings.ts
@@ -4,9 +4,7 @@ import { login_page } from "./pages/login/LoginPage.strings";
 import { register_page } from "./pages/register/RegisterPage.strings";
 import { verify_email_page } from "./pages/verify-email/VerifyEmailPage.strings";
 
-import type { I18nMessages } from "@saflib/vue-spa";
-
-export const strings: I18nMessages = {
+export const strings = {
   change_forgotten_password_page,
   forgot_password_page,
   login_page,
