+++ vue-spa/package.json
@@ -3,13 +3,13 @@
   "description": "All-in-one set of helper methods and documents for creating Vue SPAs for SAF applications",
   "type": "module",
   "exports": {
-    ".": "./index.ts",
-    "./test-utils": "./test-utils.ts",
+    ".": "./src/index.ts",
+    "./test-utils": "./src/test-utils.ts",
     "./tsconfig.app.json": "./tsconfig.app.json",
     "./tsconfig.json": "./tsconfig.json",
     "./tsconfig.node.json": "./tsconfig.node.json",
     "./workflows": "./workflows/index.ts",
-    "./tanstack": "./tanstack.ts",
+    "./tanstack": "./src/tanstack.ts",
     "./testing": "./testing/index.ts",
     "./vitest-config": "./testing/vitest-config.js"
   },
