+++ vue-spa/src/AsyncPage.vue
