+++ vue-spa/src/SpaLink.vue
