+++ vue-spa/src/assets.d.ts
