+++ vue-spa/src/events.ts
