+++ vue-spa/src/index.ts
@@ -1,0 +1,8 @@
+export { AsyncPage } from "./tricky-imports.ts";
+export * from "./types.ts";
+export { default as SpaLink } from "./SpaLink.vue";
+export * from "./tanstack.ts";
+export * from "./events.ts";
+import "./assets.d.ts";
+export * from "./app.ts";
+export * from "./strings.ts";
