+++ vue-spa/src/tricky-imports.ts
