+++ vue-spa/src/types.ts
