+++ vue-spa/workflows/spa-template/pages/home-page/HomePage.test.ts
@@ -8,7 +8,7 @@ import { type VueWrapper } from "@vue/test-utils";
 import HomePageAsync from "./HomePageAsync.vue";
 import { home_page as strings } from "./HomePage.strings.ts";
 import { router } from "../../router.ts";
-import { getElementByString } from "@saflib/vue-spa/test-utils";
+import { getElementByString } from "../../../../src/test-utils.ts";
 import type { Component } from "vue";
 import { http, HttpResponse } from "msw";
 import type { AuthResponse } from "@saflib/identity-spec"; // TODO: import the appropriate spec
